+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.4
+
+- Added SpecificDatabase to make DatabaseRepository interactions simpler when only using a single databaseName
+- Added GenericModelMap to hold some convenience function for loading maps.
+
 ## 0.1.3
 
 - Added FakeDatabaseRepository